Factory pattern supports multiple data vendors @mvandergrift Fixes #1
