+++ main.go
@@ -1,13 +1,9 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
 	"net/url"
 	"os"
 	"strconv"
@@ -19,8 +15,6 @@ import (
 	"github.com/mvandergrift/energy-sdk/auth"
 	"github.com/mvandergrift/energy-sdk/driver"
 	"github.com/mvandergrift/energy-sdk/healthmate"
-	"github.com/mvandergrift/energy-sdk/healthmate/measure"
-	"github.com/mvandergrift/energy-sdk/healthmate/workout"
 
 	"github.com/mvandergrift/energy-sdk/model"
 	"github.com/mvandergrift/energy-sdk/repo"
@@ -73,7 +67,7 @@ func main() {
 	for _, v := range result {
 		k, err := v.Export()
 		check("Export", err)
-		log.Println("Saving", v)
+		log.Println(v)
 
 		switch modelExport := k.(type) {
 		case model.Workout:
@@ -83,6 +77,7 @@ func main() {
 			measureRepo := repo.NewMeasureRepo(cn)
 			check("measureRepo.save", measureRepo.Save(&modelExport))
 		default:
+			panic(fmt.Sprintf("No handler for model.Export type %v", modelExport))
 		}
 	}
 }
@@ -97,13 +92,12 @@ func GetWorkouts(startDate string, endDate string, hc healthmate.Client) []model
 		payload.Set("startdateymd", startDate)
 		payload.Set("enddateymd", endDate)
 	} else {
-		last := strconv.FormatInt(time.Now().Add(48*time.Hour*-1).Unix(), 10)
-		log.Println("last", last)
+		last := strconv.FormatInt(time.Now().Add(72*time.Hour*-1).Unix(), 10)
 		payload.Set("lastupdate", last)
 	}
 
-	var result workout.Result
-	check("ProcessHealthmateRequest", ProcessHealthmateRequest(hc, payload, &result))
+	var result healthmate.WorkoutResult
+	check("ProcessHealthmateRequest", healthmate.ProcessRequest(hc, payload, &result))
 
 	retval := make([]model.Export, len(result.Body.Series))
 	for k, v := range result.Body.Series {
@@ -120,18 +114,19 @@ func GetMeasure(startDate string, endDate string, hc healthmate.Client) []model.
 	payload.Set("meastypes", "1,6,4,11")
 
 	if startDate != "" && endDate != "" {
+		// todo #10 Verify date formats for data provider @mvandergrift
 		unixStart, _ := time.Parse("2006-01-02", startDate)
 		unixEnd, _ := time.Parse("2006-01-02", endDate)
 		payload.Set("startdate", strconv.FormatInt(unixStart.Unix(), 10))
 		payload.Set("enddate", strconv.FormatInt(unixEnd.Unix(), 10))
 	} else {
-		last := strconv.FormatInt(time.Now().Add(168*time.Hour*-1).Unix(), 10)
-		log.Println("last", last)
+		last := strconv.FormatInt(time.Now().Add(72*time.Hour*-1).Unix(), 10)
 		payload.Set("lastupdate", last)
 	}
 
-	var result measure.Result
-	check("ProcessHealthmateRequest", ProcessHealthmateRequest(hc, payload, &result))
+	var result healthmate.MeasureResult
+	// todo #9 Capture and return error to caller @mvandergrift
+	check("ProcessHealthmateRequest", healthmate.ProcessRequest(hc, payload, &result))
 
 	var retval []model.Export
 	for _, group := range result.Body.Measuregrps {
@@ -143,54 +138,6 @@ func GetMeasure(startDate string, endDate string, hc healthmate.Client) []model.
 	}
 
 	return retval
-}
-
-/*
-Returns a new HTTPClient based on the Healthmate OAuth2 client & configurations. Uses
-the cachedTokenPath to load and store the access token needed for api authentication
-*/
-func NewHTTPClient(c healthmate.Client, cachedTokenPath string) (*http.Client, error) {
-	token, err := auth.LoadToken(cachedTokenPath)
-	if err != nil {
-		return nil, err
-	}
-
-	tokenSource := auth.RefreshToken(token, c.OAuth2Config, cachedTokenPath)
-	client := oauth2.NewClient(context.Background(), *tokenSource)
-	return client, nil
-}
-
-func ProcessHealthmateRequest(hc healthmate.Client, payload url.Values, v interface{}) error {
-	client, err := NewHTTPClient(hc, "withing.json")
-	if err != nil {
-		return fmt.Errorf("NewHTTPClient %w", err)
-	}
-
-	resp, err := client.PostForm("https://wbsapi.withings.net/v2/measure", payload)
-	if err != nil {
-		return fmt.Errorf("PostForm %w", err)
-	}
-
-	defer resp.Request.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("ioutil.ReadAll %w", err)
-	}
-
-	if *debugFlag {
-		err = ioutil.WriteFile("debug.json", body, 0644)
-		if err != nil {
-			return fmt.Errorf("WriteFile (debug) %w", err)
-		}
-	}
-
-	err = json.Unmarshal(body, v)
-	if err != nil {
-		return fmt.Errorf("Unmarshal %w", err)
-	}
-
-	return nil
 }
 
 func check(subject string, err error) {
